requirements.txt
aiocfscrape
aiohttp
async-timeout
attrs
beautifulsoup4
bitarray
certifi
chardet
cloudscraper
dnspython
future
html5lib
idna
Js2Py
msedge-selenium-tools
multidict
numpy
pandas
pdfreader
Pillow
probableparsing
pyjsparser
pymodm
pymongo
pyparsing
python-crfsuite
python-dateutil
pytz
requests
requests-toolbelt
selenium
six
soupsieve
tqdm
tzlocal
urllib3
usaddress
xlrd
yarl
